requirements.txt
mmcv-full==1.3.14
mmdet==2.14.0
mmsegmentation==0.14.1
scikit-image==0.18.3
networkx>=2.2,<2.3
numba==0.48.0
numpy<1.20.0
setuptools==59.5.0
nuscenes-devkit
plyfile
tensorboard
trimesh>=2.35.39,<2.35.40
torchmetrics==0.10.2
fvcore
sympy
